+++ Assets/HUD/StartScreenTest/MenuOverlay.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2bf4de73cf08d61449244730cc97772d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
