--- Assets/HUD/StartScreenTest/PlayGame.cs
@@ -1,19 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class PlayGame : MonoBehaviour {
-    // Changelevel function
-    public void ChangeLevel()
-    {
-        SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex + 1);
-    }
-
-    // Quit game
-    public void QuitGame()
-    {
-        Debug.Log("Quit");
-        // Application.Quit();
-    }
-}
