--- Assets/HUD/StartScreenTest/PlayGame.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d2cde76e52cafeb46b907edecb32b553
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
