+++ Assets/Lego_Assets/Cutscenes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 565eacb501575ad41b1261b959574af8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
