+++ Assets/Lego_Assets/Cutscenes/Cockpit.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 32b71c4df258a9e4189c00af9186423a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
