+++ Assets/Lego_Assets/Cutscenes/Cockpit/CockpitTimeline.playable.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a197ee6e9d501a7459ba6b370a9e22b5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
