+++ Assets/Lego_Assets/Cutscenes/Desert.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a974540da1ed9ad4da5d27fe344c3f45
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
