+++ Assets/Lego_Assets/Cutscenes/Desert/DesertTimeline.playable.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9913f3bb9e308d644b9743d6d2ec1a4b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
