+++ Assets/Lego_Assets/Lego/Figuren/Paul/Materials/FacialExpressionM 1.mat
@@ -88,7 +88,7 @@ Material:
     - _Cutoff: 0.5
     - _DetailNormalMapScale: 1
     - _DstBlend: 0
-    - _Face: -2.48
+    - _Face: 0.88
     - _GlossMapScale: 1
     - _Glossiness: 0.5
     - _GlossyReflections: 1
