+++ Assets/Lego_Assets/Lego/Figuren/Paul/Materials/MutterFacialExpressionM 2.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aba4c4ff83e727741aa43622be84c8ae
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
