+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2022.1.10f1
-m_EditorVersionWithRevision: 2022.1.10f1 (9aa0f82c4f96)
+m_EditorVersion: 2022.1.11f1
+m_EditorVersionWithRevision: 2022.1.11f1 (09bebd6e9324)
