+++ UserSettings/EditorUserSettings.asset
@@ -15,16 +15,16 @@ EditorUserSettings:
       value: 5a5757560101590a5d0c0e24427b5d44434e4c7a7b7a23677f2b4565b7b5353a
       flags: 0
     RecentlyUsedSceneGuid-3:
-      value: 02070755530c0c5f58580f2048265e4413154e7d7a2c74632e79486ab2b06468
-      flags: 0
-    RecentlyUsedSceneGuid-4:
-      value: 5b555051520c085f5d5d557215270944444e41297c7f72647d7c1835bae1603a
-      flags: 0
-    RecentlyUsedSceneGuid-5:
       value: 5b06040301540c0a545e5c7415220644124e407f7f2a7236747a1c65b4b0643d
       flags: 0
-    RecentlyUsedSceneGuid-6:
+    RecentlyUsedSceneGuid-4:
       value: 5306560050535809085c09214576094446164c29797e7e69282b1b61b6e6636c
+      flags: 0
+    RecentlyUsedSceneGuid-5:
+      value: 5b555051520c085f5d5d557215270944444e41297c7f72647d7c1835bae1603a
+      flags: 0
+    RecentlyUsedSceneGuid-6:
+      value: 02070755530c0c5f58580f2048265e4413154e7d7a2c74632e79486ab2b06468
       flags: 0
     RecentlyUsedSceneGuid-7:
       value: 5300075f030d5a0854085474417b5b444015407b752b76697a7f4467e1e4303a
